# src/main/resources/swagger/swagger.yaml
swagger: '2.0'
info:
  description: World Cup 2018 - Game Data API's
  version: 1.0.0
  title: World Cup 2018 API's
host: localhost
basePath: /game
schemes:
  - https
  - http
paths:

  '/matches':
    get:
      tags:
        - matches
      description: Find the matches
      summary: Get matches data
      responses:
        '200':
          description: Sucessfully returns the matches.
          schema:
            type: array
            items:
              $ref: '#/definitions/match'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Not found.
        '500':
          description: Internal Error.
        '504':
          description: Gateway Timeout.

  '/matches/{id}':
    get:
      tags:
        - matches
      description: Find the match by id
      summary: Get match data by id
      parameters:
        - name: id
          in: path
          description: match name
          required: true
          type: string
      responses:
        '200':
          description: Sucessfully returns the match.
          schema:
            $ref: '#/definitions/match'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Not found.
        '500':
          description: Internal Error.
        '504':
          description: Gateway Timeout.

    put:
      tags:
        - matches
      description: It updated the match result
      summary: Updated the match result
      parameters:
        - name: id
          in: path
          description: match name
          required: true
          type: string
        - name: result
          in: body
          description: Match Result data in JSON notation.
          required: true
          schema:
            $ref: '#/definitions/match_result'
      responses:
        '200':
          description: Match was updated sucessfully.
        '400':
          description: Bad Request.
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '412':
          description: Precondition failed.
          schema:
            type: array
            items:
              $ref: '#/definitions/_errorMessage'
        '422':
          description: Unprocessable Entity.
          schema:
            $ref: '#/definitions/_errorMessage'
        '500':
          description: Internal Error.
        '504':
          description: Gateway Timeout.

  '/stadiums/{id}':
    get:
      tags:
        - stadiums
      description: Find the stadium by id
      summary: Get stadium data by id
      parameters:
        - name: id
          in: path
          description: stadium id
          required: true
          type: string
      responses:
        '200':
          description: Sucessfully returns the stadium.
          schema:
            $ref: '#/definitions/stadium'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Not found.
        '500':
          description: Internal Error.
        '504':
          description: Gateway Timeout.
  '/teams/{id}':
    get:
      tags:
        - teams
      description: Find the team by id
      summary: Get team data by id
      parameters:
        - name: id
          in: path
          description: team id
          required: true
          type: string
      responses:
        '200':
          description: Sucessfully returns the team.
          schema:
            $ref: '#/definitions/team'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Not found.
        '500':
          description: Internal Error.
        '504':
          description: Gateway Timeout.

definitions:

  _errorMessage:
    type: object
    properties:
      code:
        type: string
        description: Error code.
      message:
        type: string
        description: Error message.

  stadium:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      city:
        type: string
      lat:
        type: integer
      lng:
        type: integer
      image:
        type: string

  team:
    type : object
    properties:
      id:
        type: string
      name:
        type: string
      fifaCode:
        type: string
      iso2:
        type: string
      flag:
        type: string
      emoji:
        type: string
      emojiString:
        type: string

  match:
    type : object
    properties:
      name:
        type: string
      type:
        type: string
      home_team:
        type: object
        $ref: '#/definitions/team'
      away_team:
        type: object
        $ref: '#/definitions/team'
      home_result:
        type: string
      away_result:
        type: string
      round:
        type: string
      stadium:
        type: string
      date:
        type: string
      finished:
        type: boolean

  match_result:
    type : object
    properties:
      home_result:
        type: string
      away_result:
        type: string